+++ hydromt/data_catalog.py
@@ -238,12 +238,15 @@ class DataCatalog(object):
 
         .. code-block:: console
 
-            root: <path>
-            category: <category>
+            meta:
+              root: <path>
+              category: <category>
+              version: <version>
             <name>:
               path: <path>
               data_type: <data_type>
               driver: <driver>
+              filesystem: <filesystem>
               kwargs:
                 <key>: <value>
               crs: <crs>
@@ -281,7 +284,10 @@ class DataCatalog(object):
         if root is None:
             root = meta.get("root", dirname(urlpath))
         self.from_dict(
-            yml, root=root, category=meta.get("category", None), mark_used=mark_used
+            yml,
+            root=root,
+            category=meta.get("category", None),
+            mark_used=mark_used,
         )
 
     def from_dict(
@@ -317,6 +323,7 @@ class DataCatalog(object):
                     "path": <path>,
                     "data_type": <data_type>,
                     "driver": <driver>,
+                    "filesystem": <filesystem>,
                     "kwargs": {<key>: <value>},
                     "crs": <crs>,
                     "nodata": <nodata>,
@@ -332,7 +339,11 @@ class DataCatalog(object):
             }
 
         """
-        data_dict = _parse_data_dict(data_dict, root=root, category=category)
+        data_dict = _parse_data_dict(
+            data_dict,
+            root=root,
+            category=category,
+        )
         self.update(**data_dict)
         if mark_used:
             self._used_data.extend(list(data_dict.keys()))
@@ -498,6 +509,7 @@ class DataCatalog(object):
                     source.unit_add = unit_add
                 source.path = fn_out
                 source.driver = driver
+                source.filesystem = "local"
                 source.kwargs = {}
                 source.rename = {}
                 sources_out[key] = source
@@ -754,7 +766,9 @@ class DataCatalog(object):
 
 
 def _parse_data_dict(
-    data_dict: Dict, root: Union[Path, str] = None, category: str = None
+    data_dict: Dict,
+    root: Union[Path, str] = None,
+    category: str = None,
 ) -> Dict:
     """Parse data source dictionary."""
     # link yml keys to adapter classes
@@ -788,7 +802,11 @@ def _parse_data_dict(
         elif data_type not in ADAPTERS:
             raise ValueError(f"{name}: Data type {data_type} unknown")
         adapter = ADAPTERS.get(data_type)
-        path = abs_path(root, source.pop("path"))
+        # Only for local files
+        path = source.pop("path")
+        # if remote path, keep as is else call abs_path method to solve local files
+        if not _uri_validator(path):
+            path = abs_path(root, path)
         meta = source.pop("meta", {})
         if "category" not in meta and category is not None:
             meta.update(category=category)