+++ tests/test_data_adapter.py
@@ -8,12 +8,11 @@ import geopandas as gpd
 import pandas as pd
 import xarray as xr
 import hydromt
-
-from hydromt.data_catalog import (
-    DataCatalog,
-)
+from hydromt import _compat as compat
+from hydromt.data_catalog import DataCatalog
 
 TESTDATADIR = join(dirname(abspath(__file__)), "data")
+CATALOGDIR = join(dirname(abspath(__file__)), "..", "data", "catalogs")
 
 
 def test_resolve_path(tmpdir):
@@ -63,6 +62,39 @@ def test_rasterdataset(rioda, tmpdir):
     assert np.all(da1 == rioda)
     with pytest.raises(FileNotFoundError, match="No such file or catalog key"):
         data_catalog.get_rasterdataset("no_file.tif")
+
+
+@pytest.mark.skipif(not compat.HAS_GCSFS, reason="GCSFS not installed.")
+def test_gcs_cmip6(tmpdir):
+    # TODO switch to pre-defined catalogs when pushed to main
+    catalog_fn = join(CATALOGDIR, "gcs_cmip6_data.yml")
+    data_catalog = DataCatalog(data_libs=[catalog_fn])
+    ds = data_catalog.get_rasterdataset(
+        "cmip6_NOAA-GFDL/GFDL-ESM4_historical_r1i1p1f1_Amon",
+        variables=["precip", "temp"],
+        time_tuple=(("1990-01-01", "1990-06-01")),
+    )
+    fn_nc = str(tmpdir.join("test.nc"))
+    ds.to_netcdf(fn_nc)
+    # Check reading and some preprocess
+    assert "precip" in ds
+    assert not np.any(ds[ds.raster.x_dim] > 180)
+    # Write and compare
+    ds1 = data_catalog.get_rasterdataset(fn_nc)
+    assert np.allclose(ds["precip"][0, :, :], ds1["precip"][0, :, :])
+
+
+@pytest.mark.skipif(not compat.HAS_S3FS, reason="S3FS not installed.")
+def test_aws_copdem(tmpdir):
+    # TODO switch to pre-defined catalogs when pushed to main
+    catalog_fn = join(CATALOGDIR, "aws_data.yml")
+    data_catalog = DataCatalog(data_libs=[catalog_fn])
+    da = data_catalog.get_rasterdataset(
+        "esa_worldcover_2020_v100",
+        bbox=[12.0, 46.0, 12.5, 46.50],
+    )
+    assert da.name == "landuse"
+    assert da.max().values == 100
 
 
 def test_geodataset(geoda, geodf, ts, tmpdir):